pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dvc[s3]>=3.59.2",
    "jupyter>=1.1.1",
    "keras>=3.9.2",
    "lightgbm>=4.6.0",
    "mlflow>=2.21.3",
    "pandas>=2.2.3",
    "pytorch-tabnet>=4.1.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "xgboost>=3.0.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
